GenerateScriptCommandHandler
